+++ core/upload.php
@@ -41,14 +41,19 @@ function aabs_upload($rom, $short_device, $device, $file_match, $type) {
 	$output_name = trim(shell_exec("/bin/bash -c \"basename {$output_dir}/{$file_match}\""), "\n\t");
 	$output_path = dirname("{$output_dir}/{$file_match}") . "/" . $output_name;
 
-	if (!is_file($output_path)) {
-		die("Output not found: \"{$output_path}\"\n");
+	if (AABS_IS_DRY_RUN)
+		echo "uploading build-output for '$device'\n";
+
+	if (!AABS_IS_DRY_RUN) {
+		if (!is_file($output_path)) {
+			die("Output not found: \"{$output_path}\"\n");
+		}
+
+		$build_prop = file_get_contents("{$output_dir}/system/build.prop");
+
+		$upload_dir = do_path_variables($rom, $device, $short_device, $type, AABS_UPLOAD_DIR, $build_prop);
+		$upload_file = do_path_variables($rom, $device, $short_device, $type, AABS_UPLOAD_FILE, $build_prop);
 	}
-
-	$build_prop = file_get_contents("{$output_dir}/system/build.prop");
-
-	$upload_dir = do_path_variables($rom, $device, $short_device, $type, AABS_UPLOAD_DIR, $build_prop);
-	$upload_file = do_path_variables($rom, $device, $short_device, $type, AABS_UPLOAD_FILE, $build_prop);
 
 	$hash_methods = explode(",", AABS_HASH_METHODS);
 	$hash_files = array( );
@@ -57,12 +62,20 @@ function aabs_upload($rom, $short_device, $device, $file_match, $type) {
 		$hash_method = trim($hash_method, " ");
 		$hash_path   = "{$output_dir}/{$output_name}.aabs.{$hash_method}sum";
 
-		echo "Generating {$hash_method}sum...\n";
-		$out_hash = hash_file($hash_method, $output_path);
-		file_put_contents($hash_path, "{$out_hash}  {$upload_file}");
+		if (AABS_IS_DRY_RUN)
+			echo " - generating {$hash_method}sum\n";
 
-		$hash_files[$hash_method] = $hash_path;
+		if (!AABS_IS_DRY_RUN) {
+			echo "Generating {$hash_method}sum...\n";
+			$out_hash = hash_file($hash_method, $output_path);
+			file_put_contents($hash_path, "{$out_hash}  {$upload_file}");
+
+			$hash_files[$hash_method] = $hash_path;
+		}
 	}
+
+	if (AABS_IS_DRY_RUN)
+		return;
 
 	$fn = "";
 	$params = array( );
